tools/linkedin-scanner/requirements.txt
anthropic
beautifulsoup4
python-dotenv
selenium
termcolor
webdriver-manager
pandas
tqdm
